x/eightball/types/messages_signed_scripture_list_test.go
package types

import (
	"testing"

	"eightball/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateSignedScriptureList_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateSignedScriptureList
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateSignedScriptureList{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateSignedScriptureList{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
